# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: kotoji Control-Plane API
  version: 1.0.0
  description: >
    The Go <-> Next REST/JSON control-plane API for kotoji. This spec is the SOURCE OF TRUTH
    for the Go <-> TS type sync: Go types are generated via oapi-codegen (-generate types) and
    the TS client via openapi-typescript + openapi-fetch. A CI drift gate fails on staleness.

    All schemas match the canonical Go core types frozen in CANONICAL.md. Wire field names are
    fixed there (e.g. baseSha, expected, actual, changedPaths). The data plane (static serving
    of *.hosting.example.com) is NOT part of this API and is specified in routing-and-serving.md.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: https://hosting.example.com
    description: Production control host (path-based split behind the proxy)
  - url: http://kotoji.localhost:8080
    description: Local dev
...
security:
  - sessionCookie: []
  - bearerPAT: []
paths:
...
  /auth/callback:
    get:
      tags: [auth]
      operationId: authCallback
      summary: OIDC callback (verifies state, sets the __Host- session cookie, 302s to next)
      security: []
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        "302":
          description: Login complete; session cookie set
          headers:
            Location: { schema: { type: string } }
            Set-Cookie: { schema: { type: string }, description: "__Host-kotoji_session" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/admin/github:
    get:
      tags: [admin]
      operationId: adminGetGitHub
      summary: Read the instance GitHub mirror config (admin only; secrets reduced to booleans)
      description: >
        Returns the EFFECTIVE GitHub mirror configuration (DB overrides env). It
        is SECRET-SAFE: the push token and webhook secret are NEVER returned —
        only `tokenSet`/`webhookSecretSet` "configured" booleans. Requires
        is_admin (401 anonymous, 403 non-admin).
      responses:
        "200":
          description: Effective GitHub mirror config (secret-safe)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GitHubAdminConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [admin]
      operationId: adminPutGitHub
      summary: Update the instance GitHub mirror config (admin only; token write-only)
      description: >
        Persists a PARTIAL update of the instance GitHub mirror config. Every
        body field is optional: an absent field is left unchanged. The token is
        WRITE-ONLY — supply it to set/rotate it; an empty or absent token KEEPS
        the stored one; set `clearToken: true` to remove it. The response is the
        same secret-safe view as GET (no secret values). Requires is_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GitHubAdminConfigUpdate" }
      responses:
        "200":
          description: Updated GitHub mirror config (secret-safe)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GitHubAdminConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/admin/domain:
    get:
      tags: [admin]
      operationId: adminGetDomain
      summary: Read the instance domain/URL config (admin only)
      description: >
        Returns the EFFECTIVE base domain + external control base URL with their
        sources and env-locked flags. Precedence is WordPress-style: the
        KOTOJI_BASE_DOMAIN / KOTOJI_CONTROL_BASE_URL env vars OVERRIDE the
        instance_settings DB values, which override values derived from the
        incoming request on a fresh install. A field whose env var is set has
        `*Locked: true` (read-only in the GUI). Requires is_admin (401 anonymous,
        403 non-admin). The values are NOT secret and are returned verbatim.
      responses:
        "200":
          description: Effective domain/URL config
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DomainAdminConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [admin]
      operationId: adminPutDomain
      summary: Update the instance domain/URL config (admin only)
      description: >
        Persists a PARTIAL update of the runtime base domain / control base URL.
        Every body field is optional: an absent field is left unchanged. A field
        whose env var is SET (locked) is REJECTED with 409 (the change is not
        silently no-op'd). Values are validated server-side: `baseDomain` must be
        a valid DNS hostname (no scheme/port/path), `controlBaseURL` an absolute
        http(s) origin URL. The response is the same view as GET. Requires
        is_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DomainAdminConfigUpdate" }
      responses:
        "200":
          description: Updated domain/URL config
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DomainAdminConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: A targeted field is locked by the environment (env overrides DB)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-kotoji_session
      description: Opaque server-side session id (host-only __Host- cookie). Browser/dashboard auth.
    bearerPAT:
      type: http
      scheme: bearer
      bearerFormat: kotoji_pat
      description: >
        Per-site personal access token "kotoji_pat_<base62>". Scope is the intersection of the
        token scopes {read,write,publish} and the creating member's role (see CANONICAL.md §6).

  # --------------------------------------------------------------- parameters
...
  responses:
...
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ValidationFailed:
      description: Validation failed (handle/path/branch/baseSha rules)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # --------------------------------------------------------------- schemas
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - validation
                - conflict
                - not_found
                - handle_taken
                - publish_conflict
                - branch_exists
                - nothing_to_commit
                - too_large
                - unsupported_media_type
                - rate_limited
                - quota_exceeded
                - internal
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              description: Code-specific structured detail (e.g. validation field, retry hint).
...
    GitHubAdminConfig:
      type: object
      description: >
        Secret-safe view of the instance GitHub mirror config (admin screen). The
        token and webhook secret are NEVER returned — only "configured" booleans.
        Values are the EFFECTIVE config (DB overrides env).
      required: [enabled, org, tokenSet, webhookSecretSet]
      properties:
        enabled:
          type: boolean
          description: Whether GitHub mirroring is enabled (DB override of env).
        org:
          type: string
          description: GitHub org/owner for created repos (empty if unset).
        tokenSet:
          type: boolean
          description: True when a push token is configured (DB or env). The token itself is never returned.
        webhookSecretSet:
          type: boolean
          description: True when a webhook HMAC secret is configured. The secret itself is never returned.
    GitHubAdminConfigUpdate:
      type: object
      description: >
        Partial update of the instance GitHub mirror config. All fields optional;
        absent fields are unchanged. The token is write-only (empty/absent keeps
        the stored one; clearToken removes it).
      properties:
        enabled: { type: boolean }
        org: { type: string }
        token:
          type: string
          description: New push PAT/app token (write-only; stored encrypted at rest). Empty/absent keeps the existing token.
        webhookSecret:
          type: string
          description: New webhook HMAC secret (write-only). Empty/absent keeps the existing secret.
        clearToken:
          type: boolean
          description: When true, remove the stored token (reverts to the env token if any).

    # ---------- admin (instance domain/URL config) ----------
    DomainAdminConfig:
      type: object
      description: >
        EFFECTIVE instance domain/URL config (admin screen). Each value is reported
        with its source ("env" | "db" | "derived") and a "locked" flag (true when
        the corresponding env var is set, making the field read-only). Values are
        NOT secret. Precedence: env OVERRIDES DB, DB overrides a per-request derived
        default.
      required:
        [baseDomain, controlBaseURL, baseDomainSource, controlBaseURLSource,
         baseDomainLocked, controlBaseURLLocked]
      properties:
        baseDomain:
          type: string
          example: hosting.example.com
          description: Effective base domain used to parse {handle}[--{branch}].<base>.
        controlBaseURL:
          type: string
          example: https://hosting.example.com
          description: Effective external URL of the control host.
        baseDomainSource:
          $ref: "#/components/schemas/DomainConfigSource"
        controlBaseURLSource:
          $ref: "#/components/schemas/DomainConfigSource"
        baseDomainLocked:
          type: boolean
          description: True when KOTOJI_BASE_DOMAIN is set (env wins; field read-only).
        controlBaseURLLocked:
          type: boolean
          description: True when KOTOJI_CONTROL_BASE_URL is set (env wins; field read-only).
    DomainConfigSource:
      type: string
      enum: [env, db, derived]
      description: >
        Where an effective domain/URL value came from: "env" (KOTOJI_* env var,
        locked), "db" (instance_settings, editable), or "derived" (from the
        incoming request on a fresh, unconfigured install).
    DomainAdminConfigUpdate:
      type: object
      description: >
        Partial update of the instance domain/URL config. All fields optional;
        absent fields are unchanged. A field whose env var is set is rejected with
        409 (locked by environment). An empty string reverts that field to the
        env/derived fallback (deletes the DB key).
      properties:
        baseDomain:
          type: string
          description: New base domain (valid DNS hostname; empty reverts to env/derived).
        controlBaseURL:
          type: string
          description: New control base URL (absolute http(s) origin; empty reverts to env/derived).

    # ---------- sites ----------
...